refactor(editor): renamed parent-post-body component to be consistently named compared to other components
