+++ honestcash-v2/src/app/modules/editor/components/parent-post-body/parent-post-body.component.scss
